[TEST] Remove HeadBodyIsEmptyIntegTestCase as it has only one subclass
Move all its methods to its only subclass: Netty4HeadBodyIsEmptyIT. This has to with the recent removal of the netty3 plugin from the codebase.
